python-export-unpaid-invoices-to-cashcollector-xls/requirements.txt
requests==2.32.3
python_dotenv==1.0.1
openpyxl==3.1.5
python-dateutil==2.9.0